refactored libs and files
