+++ biosimulator_processes/container-assets/pyproject.toml
@@ -1,0 +1,33 @@
+[tool.poetry]
+name = "biosimulator-processes"
+version = "0.0.16"
+description = "Core implementations of process-bigraph.composite.Process aligning with BioSimulators simulator tools."
+authors = ["Alex Patrie <alexanderpatrie@gmail.com>"]
+readme = "README.md"
+packages = [{include = "biosimulator_processes"}]
+
+[tool.poetry.dependencies]
+python = ">=3.10"
+zarr = "*"
+termcolor = "*"
+jupyterlab = "*"
+bigraph-schema = "*"
+process-bigraph = "*"
+numpy = "*" # "^1.26.4"
+copasi-basico = "*"
+cobra = "*"
+tellurium = "*"
+smoldyn = "*"
+bigraph-viz = "^0.0.31"
+python-libsbml = "^5.20.2"
+docker = ">=7.0"
+
+[tool.poetry.group.dev.dependencies]
+pytest = "*"
+mypy = ">=1.8.0"
+flake8 = "*"
+twine = "*"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
